Makefiles changed && cpp01 ex05 main changed

+++ cpp01/ex01/zombieHorde.cpp
@@ -3,7 +3,7 @@
 Zombie* zombieHorde( int N, std::string name )
 {
     Zombie *zombie = new Zombie[N];
-    for (size_t i = 0; i < N; i++)
+    for (int i = 0; i < N; i++)
     {
         zombie[i].setName(name);
         zombie[i].announce();
